requirements.txt
PySimpleGUI
python-vlc